doc: improving documentation as example.
typing hels to understand the code, but in python is not mandatory.
also, typing cannot limit the variable type, for example `var: int` still can be a string `var: int = "string"`.
but helps the user and developer undertand the right data type.

the docstring is a good way to help the user and developer understand the code.
and helps bring to the IDE the documentation of class, methos and functions.